Make Confal Pyramids colors and visibility adjustable and move some potentially shared code into a common unit
